Fixed neural crafting, jei crash continues.

+++ common/src/main/java/org/chubby/github/mobcontroller/api/energy/EnergyStorage.java
@@ -12,7 +12,7 @@ public interface EnergyStorage {
      * @param simulate If true, does not actually modify the storage.
      * @return The amount of energy successfully inserted.
      */
-    long insert(long amount, boolean simulate);
+    int insert(int amount, boolean simulate);
 
     /**
      * Extracts energy from the storage.
@@ -21,21 +21,21 @@ public interface EnergyStorage {
      * @param simulate If true, does not actually modify the storage.
      * @return The amount of energy successfully extracted.
      */
-    long extract(long amount, boolean simulate);
+    int extract(int amount, boolean simulate);
 
     /**
      * Gets the amount of energy currently stored.
      *
      * @return The stored energy value.
      */
-    long getStoredEnergy();
+    int getStoredEnergy();
 
     /**
      * Gets the maximum energy capacity of this storage.
      *
      * @return The maximum energy capacity.
      */
-    long getCapacity();
+    int getCapacity();
 
     /**
      * Checks if energy can be inserted into this storage.
