+++ common/src/main/java/org/chubby/github/mobcontroller/api/energy/EnergyStorageImpl.java
@@ -5,24 +5,24 @@ import net.minecraft.nbt.CompoundTag;
 import net.minecraft.util.Mth;
 
 public abstract class EnergyStorageImpl implements EnergyStorage {
-    protected long energy;
-    protected long capacity;
-    protected long maxReceive;
-    protected long maxExtract;
+    protected int energy;
+    protected int capacity;
+    protected int maxReceive;
+    protected int maxExtract;
 
-    public EnergyStorageImpl(long capacity) {
+    public EnergyStorageImpl(int capacity) {
         this(capacity, capacity, capacity, 0);
     }
 
-    public EnergyStorageImpl(long capacity, long maxTransfer) {
+    public EnergyStorageImpl(int capacity, int maxTransfer) {
         this(capacity, maxTransfer, maxTransfer, 0);
     }
 
-    public EnergyStorageImpl(long capacity, long maxReceive, long maxExtract) {
+    public EnergyStorageImpl(int capacity, int maxReceive, int maxExtract) {
         this(capacity, maxReceive, maxExtract, 0);
     }
 
-    public EnergyStorageImpl(long capacity, long maxReceive, long maxExtract, long energy) {
+    public EnergyStorageImpl(int capacity, int maxReceive, int maxExtract, int energy) {
         this.capacity = capacity;
         this.maxReceive = maxReceive;
         this.maxExtract = maxExtract;
@@ -30,12 +30,12 @@ public abstract class EnergyStorageImpl implements EnergyStorage {
     }
 
     @Override
-    public long insert(long amount, boolean simulate) {
+    public int insert(int amount, boolean simulate) {
         if (!canInsert() || amount <= 0) {
             return 0;
         }
 
-        long energyReceived = Math.min(this.capacity - this.energy, Math.min(this.maxReceive, amount));
+        int energyReceived = Math.min(this.capacity - this.energy, Math.min(this.maxReceive, amount));
         if (!simulate) {
             this.energy += energyReceived;
         }
@@ -43,12 +43,12 @@ public abstract class EnergyStorageImpl implements EnergyStorage {
     }
 
     @Override
-    public long extract(long amount, boolean simulate) {
+    public int extract(int amount, boolean simulate) {
         if (!canExtract() || amount <= 0) {
             return 0;
         }
 
-        long energyExtracted = Math.min(this.energy, Math.min(this.maxExtract, amount));
+        int energyExtracted = Math.min(this.energy, Math.min(this.maxExtract, amount));
         if (!simulate) {
             this.energy -= energyExtracted;
         }
@@ -56,12 +56,12 @@ public abstract class EnergyStorageImpl implements EnergyStorage {
     }
 
     @Override
-    public long getStoredEnergy() {
+    public int getStoredEnergy() {
         return this.energy;
     }
 
     @Override
-    public long getCapacity() {
+    public int getCapacity() {
         return this.capacity;
     }
 
@@ -75,24 +75,24 @@ public abstract class EnergyStorageImpl implements EnergyStorage {
         return this.maxReceive > 0;
     }
 
-    public long getEnergy() {
+    public int getEnergy() {
         return energy;
     }
 
-    public long getMaxReceive() {
+    public int getMaxReceive() {
         return maxReceive;
     }
 
-    public long getMaxExtract() {
+    public int getMaxExtract() {
         return maxExtract;
     }
 
     public void save(CompoundTag tag, HolderLookup.Provider registries) {
-        tag.putLong("EnergyStored", this.energy);
+        tag.putInt("EnergyStored", this.energy);
     }
 
     public void load(CompoundTag tag, HolderLookup.Provider registries) {
-        this.energy = tag.getLong("EnergyStored");
+        this.energy = tag.getInt(("EnergyStored"));
     }
 
     public abstract void onChange();
